Shots are saving to the database.... only for test game.

+++ Cellular/ViewModel/Shot.cs
@@ -16,7 +16,7 @@ namespace Cellular.ViewModel
         public int? ShotNumber { get; set; }
         public int? Ball { get; set; }
         public int? Count { get; set; }
-        public byte? LeaveType { get; set; }
+        public short? LeaveType { get; set; }
         public string? Side { get; set; }
         public string? Position { get; set; }
